+++ src/sql/__init__.py
@@ -1,3 +1,10 @@
-from .magic import *
+from .magic import RenderMagic, SqlMagic, load_ipython_extension
 
-__version__ = "0.4.7dev"
+__version__ = "0.5.2dev"
+
+
+__all__ = [
+    "RenderMagic",
+    "SqlMagic",
+    "load_ipython_extension",
+]