+++ src/sql/connection.py
@@ -1,7 +1,7 @@
 import os
-import re
 
 import sqlalchemy
+from sqlalchemy.engine import Engine
 
 
 class ConnectionError(Exception):
@@ -23,50 +23,99 @@ def rough_dict_get(dct, sought, default=None):
     return default
 
 
-class Connection(object):
+class Connection:
+    """Manages connections to databases
+
+    Parameters
+    ----------
+    engine: sqlalchemy.engine.Engine
+        The SQLAlchemy engine to use
+    """
+
+    # the active connection
     current = None
+
+    # all connections
     connections = {}
 
     @classmethod
     def tell_format(cls):
+        """
+        Returns an error message that we can display to the user
+        to tell them how to pass the connection string
+        """
         return """Connection info needed in SQLAlchemy format, example:
                postgresql://username:password@hostname/dbname
                or an existing connection: %s""" % str(
             cls.connections.keys()
         )
 
-    def __init__(self, connect_str=None, connect_args={}, creator=None):
-        try:
-            if creator:
-                engine = sqlalchemy.create_engine(
-                    connect_str, connect_args=connect_args, creator=creator
-                )
-            else:
-                engine = sqlalchemy.create_engine(
-                    connect_str, connect_args=connect_args
-                )
-        except:  # TODO: bare except; but what's an ArgumentError?
-            print(self.tell_format())
-            raise
+    def __init__(self, engine, alias=None):
         self.dialect = engine.url.get_dialect()
         self.metadata = sqlalchemy.MetaData(bind=engine)
         self.name = self.assign_name(engine)
         self.session = engine.connect()
-        self.connections[repr(self.metadata.bind.url)] = self
-        self.connect_args = connect_args
+        self.connections[alias or repr(self.metadata.bind.url)] = self
+        self.connect_args = None
+        self.alias = alias
         Connection.current = self
 
     @classmethod
-    def set(cls, descriptor, displaycon, connect_args={}, creator=None):
-        "Sets the current database connection"
+    def from_connect_str(
+        cls, connect_str=None, connect_args=None, creator=None, alias=None
+    ):
+        """Creates a new connection from a connection string"""
+        connect_args = connect_args or {}
+
+        try:
+            if creator:
+                engine = sqlalchemy.create_engine(
+                    connect_str,
+                    connect_args=connect_args,
+                    creator=creator,
+                )
+            else:
+                engine = sqlalchemy.create_engine(
+                    connect_str,
+                    connect_args=connect_args,
+                )
+        except Exception:
+            print(cls.tell_format())
+            raise
+
+        connection = cls(engine, alias=alias)
+        connection.connect_args = connect_args
+
+        return connection
+
+    @classmethod
+    def set(cls, descriptor, displaycon, connect_args=None, creator=None, alias=None):
+        """
+        Sets the current database connection
+        """
+        connect_args = connect_args or connect_args
 
         if descriptor:
             if isinstance(descriptor, Connection):
                 cls.current = descriptor
+            elif isinstance(descriptor, Engine):
+                cls.current = Connection(descriptor)
             else:
                 existing = rough_dict_get(cls.connections, descriptor)
-            # http://docs.sqlalchemy.org/en/rel_0_9/core/engines.html#custom-dbapi-connect-arguments
-            cls.current = existing or Connection(descriptor, connect_args, creator)
+
+                # NOTE: I added one indentation level, otherwise
+                # the "existing" variable would not exist if
+                # passing an engine object as descriptor.
+                # Since I never saw this breaking, my guess
+                # is that we're missing some unit tests
+                # when descriptor is a connection object
+                # http://docs.sqlalchemy.org/en/rel_0_9/core/engines.html#custom-dbapi-connect-arguments # noqa
+                cls.current = existing or Connection.from_connect_str(
+                    connect_str=descriptor,
+                    connect_args=connect_args,
+                    creator=creator,
+                    alias=alias,
+                )
         else:
 
             if cls.connections:
@@ -74,12 +123,16 @@ class Connection(object):
                     print(cls.connection_list())
             else:
                 if os.getenv("DATABASE_URL"):
-                    cls.current = Connection(
-                        os.getenv("DATABASE_URL"), connect_args, creator
+                    cls.current = Connection.from_connect_str(
+                        connect_str=os.getenv("DATABASE_URL"),
+                        connect_args=connect_args,
+                        creator=creator,
+                        alias=alias,
                     )
                 else:
                     raise ConnectionError(
-                        "Environment variable $DATABASE_URL not set, and no connect string given."
+                        "Environment variable $DATABASE_URL "
+                        "not set, and no connect string given."
                     )
         return cls.current
 
@@ -92,14 +145,18 @@ class Connection(object):
     def connection_list(cls):
         result = []
         for key in sorted(cls.connections):
-            engine_url = cls.connections[
-                key
-            ].metadata.bind.url  # type: sqlalchemy.engine.url.URL
-            if cls.connections[key] == cls.current:
-                template = " * {}"
+            conn = cls.connections[key]
+            engine_url = conn.metadata.bind.url  # type: sqlalchemy.engine.url.URL
+
+            prefix = "* " if conn == cls.current else "  "
+
+            if conn.alias:
+                repr_ = f"{prefix} ({conn.alias}) {engine_url!r}"
             else:
-                template = "   {}"
-            result.append(template.format(engine_url.__repr__()))
+                repr_ = f"{prefix} {engine_url!r}"
+
+            result.append(repr_)
+
         return "\n".join(result)
 
     @classmethod
@@ -115,7 +172,12 @@ class Connection(object):
                 "Could not close connection because it was not found amongst these: %s"
                 % str(cls.connections.keys())
             )
-        cls.connections.pop(str(conn.metadata.bind.url))
+
+        if descriptor in cls.connections:
+            cls.connections.pop(descriptor)
+        else:
+            cls.connections.pop(str(conn.metadata.bind.url))
+
         conn.session.close()
 
     def close(self):